refactor: errorContext를 DataContext에서 관리하도록 통합하고 일부 로직을 핸들러로 분리
