+++ src/context/DataContext.tsx
@@ -10,6 +10,8 @@ interface DataContextType {
   isLoading: Map<string, boolean>;
   setIsLoading: React.Dispatch<SetStateAction<Map<string, boolean>>>;
   handleLoading: (dataLoading: boolean, dataName: string) => void;
+  errorMessage: string;
+  setErrorMessage: React.Dispatch<SetStateAction<string>>;
 }
 
 export const DataContext = createContext<DataContextType | null>(null);
@@ -17,6 +19,8 @@ export const DataContext = createContext<DataContextType | null>(null);
 export const DataProvider = ({ children }: DataContextProps) => {
   const [data, setData] = useState<Map<string, unknown>>(new Map());
   const [isLoading, setIsLoading] = useState<Map<string, boolean>>(new Map());
+
+  const [errorMessage, setErrorMessage] = useState('');
 
   const handleLoading = useCallback((dataLoading: boolean, dataName: string) => {
     setIsLoading((prev) => new Map(prev).set(dataName, dataLoading));
@@ -30,6 +34,8 @@ export const DataProvider = ({ children }: DataContextProps) => {
         isLoading,
         setIsLoading,
         handleLoading,
+        errorMessage,
+        setErrorMessage,
       }}
     >
       {children}
