+++ src/pages/Layout.tsx
@@ -1,19 +1,16 @@
 import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
 import Header from '../components/Header';
-import { ErrorMessageProvider } from '../context/ErrorMessageContext';
 import { DataProvider } from '../context/DataContext';
 
 const Layout = () => {
   return (
-    <ErrorMessageProvider>
-      <DataProvider>
-        <LayoutContainer>
-          <Header />
-          <Outlet />
-        </LayoutContainer>
-      </DataProvider>
-    </ErrorMessageProvider>
+    <DataProvider>
+      <LayoutContainer>
+        <Header />
+        <Outlet />
+      </LayoutContainer>
+    </DataProvider>
   );
 };
